# relayer/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Bosphor Relayer API
  version: 0.1.0
  summary: Public HTTP API of the Bosphor relayer.
  description: |
    Bosphor routes a storage intent from an origin chain (EVM or Solana) to
    Walrus on Sui and brings a storage proof back. The relayer is the off-chain
    party that prices the store, receives the file bytes out-of-band, stores them
    on Walrus and reports progress.

    A typical integration calls, in order:

    1. `POST /quote` to price the store in the origin chain's native token. On
       Solana, first read the live LayerZero fee from `GET /lz-fee/solana` and
       pass it as `forwardLzFeeNative`.
    2. `POST /blob/encode` (optional) to derive the Walrus blob id to commit to.
    3. `submitIntent` on the origin chain adapter (not part of this API).
    4. `POST /blob/{intentId}` to upload the exact bytes you committed to.
    5. `GET /public/intents` to follow the intent across its hops.

    The `@bosphor/sdk` package wraps all of this. Use this reference when you
    call the relayer directly or build your own client.

    **Amounts.** Native token amounts are decimal strings in the smallest unit
    (wei for ETH, lamports for SOL) so no precision is lost in JSON. USD values
    in a quote breakdown are plain JSON numbers.

    **Errors.** Every error response is JSON of the shape
    `{ "statusCode": 400, "message": "...", "error": "Bad Request" }`.

...
    `X-RateLimit-Limit` and `X-RateLimit-Remaining` headers.

...
  license:
...
servers:
...
paths:
...
  /blob/encode:
    post:
      tags: [Ingest]
      operationId: encodeBlob
      summary: Derive a Walrus blob id
      description: |
        Send raw bytes, get back the Walrus blob id you must commit to in
        `submitIntent`. Nothing is stored and no WAL is spent. This lets a browser
        dApp derive the blob id without bundling the Walrus SDK or depending on a
        public Walrus publisher.

        This route has its own, tighter per-IP rate limit because encoding is CPU
        heavy.
      parameters:
        - $ref: '#/components/parameters/AppId'
      requestBody:
        $ref: '#/components/requestBodies/RawBlob'
      responses:
        '201':
          description: The derived blob id.
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncodeResponse'
              examples:
                encoded:
                  summary: A 35 byte blob
                  value:
                    blobId: mLzOEj8l3vxE6CbAvIvtNMeDxOD5DQmvK08kSpFGLp0
                    size: 35
        '400':
          $ref: '#/components/responses/BadRawBody'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
    AppId:
      name: X-Bosphor-App
      in: header
      required: false
      description: |
        Optional short name for your application, for example `my-dapp`. It is
        recorded with your intents so usage from your app is counted separately,
        and it can get its own rate-limit budget. It is attribution only, not
        authentication: anyone can send any value and it grants nothing. Ids are
        case-insensitive. A malformed value is rejected with `400`.
      schema:
        type: string
        pattern: '^[a-zA-Z0-9][a-zA-Z0-9\-_.]{0,63}$'
        maxLength: 64
      example: my-dapp
  headers:
    Retry-After:
      description: Seconds to wait before retrying.
      schema:
        type: integer
        minimum: 1
      example: 42
    X-RateLimit-Limit:
      description: Requests allowed in the current window for the tightest budget that applies.
      schema:
        type: integer
      example: 120
    X-RateLimit-Remaining:
      description: Requests left in the current window for the tightest budget that applies.
      schema:
        type: integer
      example: 119
  requestBodies:
    RawBlob:
      required: true
      description: |
        The raw blob bytes. The relayer accepts up to 10 MiB (10485760 bytes) per
        blob by default; an operator can change this with `MAX_INGEST_BLOB_BYTES`.
      content:
        application/octet-stream:
          schema:
            type: string
            contentMediaType: application/octet-stream
            maxLength: 10485760
  responses:
    BadRawBody:
      description: The body is empty, or `X-Bosphor-App` is malformed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            emptyBody:
              summary: No bytes in the body
              value:
                statusCode: 400
                message: request body must be the raw blob bytes
                error: Bad Request
            badAppId:
              $ref: '#/components/examples/BadAppId'
    PayloadTooLarge:
      description: |
        The body is larger than the relayer's ingest cap (default 10 MiB). A body
        far above the cap may be refused before it is read, in which case the
        response body may not be JSON.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 413
            message: blob exceeds MAX_INGEST_BLOB_BYTES (10485761 > 10485760)
            error: Payload Too Large
    TooManyRequests:
      description: |
        A rate-limit budget is exhausted. Wait for the number of seconds in
        `Retry-After` and try again. The message names the budget that tripped:
        `ip`, `encode` or `app`.
      headers:
        Retry-After:
          $ref: '#/components/headers/Retry-After'
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          description: Always `0` on a rate-limited response.
          schema:
            type: integer
            const: 0
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 429
            error: Too Many Requests
            message: rate limit exceeded (ip); retry after 42s
    InternalError:
      description: |
        Unexpected failure, for example no usable price source for a quote. A
        malformed request is always a `400`, never a `500`. The body carries no
        `error` field.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 500
            message: Internal server error
  examples:
    BadAppId:
      summary: Malformed X-Bosphor-App header (hosted testnet)
      value:
        message: 'invalid X-Bosphor-App header: expected a slug matching /^[a-z0-9][a-z0-9\-_.]{0,63}$/i'
        error: Bad Request
        statusCode: 400
  schemas:
    Error:
      type: object
      description: Error body returned by every failing request.
      required: [statusCode, message]
      properties:
        statusCode:
          type: integer
          description: The HTTP status code, repeated in the body.
        message:
          description: Human-readable detail. Do not parse it; branch on `statusCode`.
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        error:
          type: string
          description: The HTTP status text, for example `Bad Request`. Absent on `500`.
...
    BlobId:
      type: string
      description: Walrus blob id, base64url without padding (as shown on Walruscan).
      examples: [mLzOEj8l3vxE6CbAvIvtNMeDxOD5DQmvK08kSpFGLp0]
    EncodeResponse:
      type: object
      additionalProperties: false
      required: [blobId, size]
      properties:
        blobId:
          $ref: '#/components/schemas/BlobId'
        size:
          type: integer
          minimum: 1
          description: Byte length of the body.
...